core/requirements_test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements_test.txt requirements_test.in
#
aiodnsresolver==0.0.120
aiohttp==3.5.4
aioredis==1.1.0
aspy.yaml==1.2.0          # via pre-commit
astroid==2.2.5            # via pylint
async-timeout==3.0.0      # via aiohttp, aioredis
attrs==18.1.0             # via aiohttp
cfgv==1.5.0               # via pre-commit
chardet==3.0.4            # via aiohttp
coverage==4.5.1
freezegun==0.3.10
hawk-server-asyncio==0.0.1
hiredis==0.2.0            # via aioredis
identify==1.4.0           # via pre-commit
idna==2.7                 # via yarl
importlib-metadata==0.8   # via pre-commit
isort==4.3.4              # via pylint
lazy-object-proxy==1.3.1  # via astroid
mccabe==0.6.1             # via pylint
mohawk==0.3.4
multidict==4.3.1          # via aiohttp, yarl
nodeenv==1.3.3            # via pre-commit
pre-commit==1.14.4
prometheus_client==0.3.0
pylint==2.3.1
python-dateutil==2.7.3    # via freezegun
pyyaml==5.1               # via aspy.yaml, pre-commit
raven==6.9.0
six==1.11.0               # via astroid, cfgv, freezegun, mohawk, pre-commit, python-dateutil
toml==0.10.0              # via pre-commit
typed-ast==1.3.1          # via astroid
virtualenv==16.4.3        # via pre-commit
wrapt==1.10.11            # via astroid
yarl==1.2.6               # via aiohttp
zipp==0.3.3               # via importlib-metadata
